refactor era5 source domain module
